@@ -1,31 +1,45 @@
 {
   "name": "frontend",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@emotion/react": "^11.10.0",
+    "@emotion/styled": "^11.10.0",
+    "@mui/icons-material": "^5.10.2",
+    "@mui/material": "^5.10.0",
+    "@reduxjs/toolkit": "^1.6.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.2.0",
+    "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
+    "bootstrap": "^5.2.0-beta1",
+    "framer-motion": "^6.3.4",
+    "http-proxy-middleware": "^2.0.0",
+    "lodash": "^4.17.21",
+    "react": "^18.1.0",
+    "react-apexcharts": "^1.4.0",
+    "react-bootstrap": "^2.3.1",
+    "react-bootstrap-icons": "^1.6.1",
+    "react-chartjs-2": "^4.1.0",
+    "react-dom": "^18.1.0",
+    "react-helmet": "^6.1.0",
+    "react-icons": "^4.4.0",
+    "react-intl": "^6.1.2",
+    "react-scripts": "5.0.1",
+    "react-router-dom": "^5.2.0",
+    "web-vitals": "^2.1.4"
+  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "jest",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@reduxjs/toolkit": "^1.6.2",
-    "axios": "^0.21.1",
-    "bootstrap": "^5.2.2",
-    "framer-motion": "5.3.0",
-    "http-proxy-middleware": "^2.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.0.0-beta.6",
-    "react-dom": "^17.0.2",
-    "react-helmet": "^6.1.0",
-    "react-intl": "^6.1.2",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^2.1.3",
-    "web-vitals": "^3.0.3"
+  "eslintConfig": {
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
